CMake build system is updated. Build shared libs only (by default).
We are building now ACADO ONLY the shared lib (by default) on Unix like
OSs, and static for MSVC. All other libs are built as static and linked
against the main ACADO lib. That way, we only have to ship one library. 

Building of the static ACADO lib is optional, just set
ACADO_BUILD_STATIC variable to ON.

Building of the shared libs for Windows is disabled.

qpOASES 3.0 build script is heavily modified, just to satisfy needs for
building ACADO.