Bugfix: allow compilation with __DEBUG__.
The argument idx_ is only used in an ASSERT; ASSERT, in turn, is a
no-op unless __DEBUG__ is defined.  So, when __DEBUG__ is not defined,
gcc would warn about an unused variable if the name is declared.

Allow debug and warning-free non-debug compilation to occur by making
argument name declaration conditional on __DEBUG__.